remove deprecated np.int & np.float usage
Newer versions of numpy deprecated these aliases for the base `int` and
`float` data types.

Removing usage of them in the codebase.

+++ src/wepy/resampling/resamplers/revo.py
@@ -115,9 +115,9 @@ class REVOResampler(CloneMergeResampler):
         (1,),
     )
     RESAMPLER_DTYPES = CloneMergeResampler.RESAMPLER_DTYPES + (
-        np.int,
-        np.float,
-        np.float,
+        int,
+        float,
+        float,
     )
 
     # fields that can be used for a table like representation
